Move behavior tests to integrationTest directory
