+++ richtextfx/build.gradle
@@ -1,20 +1,39 @@
+buildscript {
+    repositories {
+        jcenter()
+    }
+    dependencies {
+        classpath 'org.unbroken-dome.gradle-plugins:gradle-testsets-plugin:1.3.2'
+    }
+}
+
 apply plugin: 'maven'
 apply plugin: 'signing'
+apply plugin: 'org.unbroken-dome.test-sets'
 
 group = 'org.fxmisc.richtext'
+
+testSets {
+    integrationTest { dirName = 'integrationTest' }
+}
+
+check.dependsOn integrationTest
+integrationTest.mustRunAfter test
 
 dependencies {
     compile group: 'org.reactfx', name: 'reactfx', version: '2.0-M5'
     compile (group: 'org.fxmisc.undo', name: 'undofx', version: '1.3.1')
     compile (group: 'org.fxmisc.flowless', name: 'flowless', version: '0.5.2')
     compile group: 'org.fxmisc.wellbehaved', name: 'wellbehavedfx', version: '0.3'
+
     testCompile group: 'junit', name: 'junit', version: '[4.0,5)'
 
-    testCompile "org.testfx:testfx-core:4.0.5-alpha"
-    testCompile ("org.testfx:testfx-junit:4.0.5-alpha") {
+    integrationTestCompile group: 'junit', name: 'junit', version: '[4.0,)'
+    integrationTestCompile group: 'com.nitorcreations', name: 'junit-runners', version: '1.2'
+    integrationTestCompile "org.testfx:testfx-core:4.0.5-alpha"
+    integrationTestCompile ("org.testfx:testfx-junit:4.0.5-alpha") {
         exclude(group: "junit", module: "junit")
     }
-    testCompile group: 'com.nitorcreations', name: 'junit-runners', version: '1.2'
 }
 
 javadoc {
